Test projesi eklendi, CreateBook testleri yazıldı

+++ OrtaSeviyeDotnetCorePatikasi/02.DotnetCore/Pratikler/BookStore/WebApi/Application/AuthorOperations/Commands/DeleteAuthor/DeleteAuthorCommand.cs
@@ -7,17 +7,23 @@ namespace WebApi.Application.AuthorOperations.DeleteAuthor
     public class DeleteAuthorCommand
     {
         public int AuthorId { get; set; }
-        private readonly BookStoreDbContext _context;
+        private readonly IBookStoreDbContext _context;
 
-        public DeleteAuthorCommand(BookStoreDbContext context)
+        public DeleteAuthorCommand(IBookStoreDbContext context)
         {
             _context = context;
         }
 
-        public void Handle(){
+        public void Handle()
+        {
             var author = _context.Authors.SingleOrDefault(x => x.Id == AuthorId);
-            if(author is null){
+            if (author is null)
+            {
                 throw new InvalidOperationException("Silinecek yazar bulunamadı.");
+            }
+            if ((_context.Books.SingleOrDefault(book => book.AuthorId == author.Id)) is not null)
+            {
+                throw new InvalidOperationException("Kitabı yayında olan yazar silinemez.");
             }
             _context.Authors.Remove(author);
             _context.SaveChanges();
